refactor: split analysis jobs db module
