+++ src/egui_app/controller/analysis_jobs/db/constants.rs
@@ -1,0 +1,4 @@
+pub(super) const ANALYZE_SAMPLE_JOB_TYPE: &str = "wav_metadata_v1";
+pub(super) const REBUILD_INDEX_JOB_TYPE: &str = "rebuild_index_v1";
+pub(super) const EMBEDDING_BACKFILL_JOB_TYPE: &str = "embedding_backfill_v1";
+pub(super) const DEFAULT_JOB_TYPE: &str = ANALYZE_SAMPLE_JOB_TYPE;
